windows version working
