--- CMakePresets.json
@@ -1,47 +1,0 @@
-{
-  "version": 4,
-  "include": ["./ConanPresets.json"],
-  "configurePresets": [
-    {
-      "name": "default",
-      "displayName": "multi config",
-      "inherits": "conan-default"
-    },
-    {
-      "name": "release",
-      "displayName": "release single config",
-      "inherits": "conan-release"
-    },
-    {
-      "name": "debug",
-      "displayName": "debug single config",
-      "inherits": "conan-debug"
-    }
-  ],
-  "buildPresets": [
-    {
-      "name": "multi-release",
-      "configurePreset": "default",
-      "configuration": "Release",
-      "inherits": "conan-release"
-    },
-    {
-      "name": "multi-debug",
-      "configurePreset": "default",
-      "configuration": "Debug",
-      "inherits": "conan-debug"
-    },
-    {
-      "name": "release",
-      "configurePreset": "release",
-      "configuration": "Release",
-      "inherits": "conan-release"
-    },
-    {
-      "name": "debug",
-      "configurePreset": "debug",
-      "configuration": "Debug",
-      "inherits": "conan-debug"
-    }
-  ]
-}
