+++ amcset_gui/CMakeLists.txt
@@ -1,24 +1,146 @@
-# cmake_minimum_required(VERSION 3.15)
+# # cmake_minimum_required(VERSION 3.15)
 
-# project(acmset_gui VERSION 0.1 LANGUAGES CXX)
+# # project(acmset_gui VERSION 0.1 LANGUAGES CXX)
 
+
+# set(TS_FILES amcset_gui_en_US.ts)
+
+# set(PROJECT_SOURCES
+#         ${PROJECT_SOURCE_DIR}/amcset_gui/src/main.cpp
+#         ${PROJECT_SOURCE_DIR}/amcset_gui/src/amcset_main_window.cpp
+#         ${PROJECT_SOURCE_DIR}/amcset_gui/src/amcset_main_window.ui
+#         ${TS_FILES}
+# )
+
+
+# if(${QT_VERSION_MAJOR} GREATER_EQUAL 6)
+#     qt_add_executable(amcset_gui
+#         MANUAL_FINALIZATION
+#         ${PROJECT_SOURCES}
+#     )
+# # Define target properties for Android with Qt 6 as:
+# #    set_property(TARGET amcset_gui APPEND PROPERTY QT_ANDROID_PACKAGE_SOURCE_DIR
+# #                 ${CMAKE_CURRENT_SOURCE_DIR}/android)
+# # For more information, see https://doc.qt.io/qt-6/qt-add-executable.html#target-creation
+
+#     qt_create_translation(QM_FILES ${CMAKE_SOURCE_DIR} ${TS_FILES})
+# else()
+#     if(ANDROID)
+#         add_library(amcset_gui SHARED
+#             ${PROJECT_SOURCES}
+#         )
+# # Define properties for Android with Qt 5 after find_package() calls as:
+# #    set(ANDROID_PACKAGE_SOURCE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/android")
+#     else()
+#         add_executable(amcset_gui
+#             ${PROJECT_SOURCES}
+#         )
+#     endif()
+
+#     qt5_create_translation(QM_FILES ${CMAKE_SOURCE_DIR} ${TS_FILES})
+# endif()
+
+# target_include_directories(amcset_gui PRIVATE
+#     ${PROJECT_SOURCE_DIR}/amcset_gui/include
+# )
+
+# target_link_libraries(amcset_gui PRIVATE Qt${QT_VERSION_MAJOR}::Widgets)
+
+# # Qt for iOS sets MACOSX_BUNDLE_GUI_IDENTIFIER automatically since Qt 6.1.
+# # If you are developing for iOS or macOS you should consider setting an
+# # explicit, fixed bundle identifier manually though.
+# if(${QT_VERSION} VERSION_LESS 6.1.0)
+#   set(BUNDLE_ID_OPTION MACOSX_BUNDLE_GUI_IDENTIFIER com.example.amcset_gui)
+# endif()
+# set_target_properties(amcset_gui PROPERTIES
+#     ${BUNDLE_ID_OPTION}
+#     MACOSX_BUNDLE_BUNDLE_VERSION ${PROJECT_VERSION}
+#     MACOSX_BUNDLE_SHORT_VERSION_STRING ${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}
+#     MACOSX_BUNDLE TRUE
+#     WIN32_EXECUTABLE TRUE
+# )
+
+# include(GNUInstallDirs)
+# install(TARGETS amcset_gui
+#     BUNDLE DESTINATION .
+#     LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+#     RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
+# )
+
+# if(QT_VERSION_MAJOR EQUAL 6)
+#     qt_finalize_executable(amcset_gui)
+# endif()
+
+set(CMAKE_AUTOUIC ON)
+set(CMAKE_AUTOMOC ON)
+set(CMAKE_AUTORCC ON)
+
+set(CMAKE_CXX_STANDARD 17)
+set(CMAKE_CXX_STANDARD_REQUIRED ON)
+
+find_package(QT NAMES Qt6 Qt5 REQUIRED COMPONENTS Widgets LinguistTools)
+find_package(Qt${QT_VERSION_MAJOR} REQUIRED COMPONENTS Widgets LinguistTools)
 
 set(TS_FILES amcset_gui_en_US.ts)
 
 set(PROJECT_SOURCES
         main.cpp
-        amcset_main_window.cpp
-        amcset_main_window.h
-        amcset_main_window.ui
+        mainwindow.cpp
+        mainwindow.h
+        mainwindow.ui
+        ${TS_FILES}
 )
 
-qt_add_executable(amcset_gui
-    ${PROJECT_SOURCES}
-)
+if(${QT_VERSION_MAJOR} GREATER_EQUAL 6)
+    qt_add_executable(amcset_gui
+        MANUAL_FINALIZATION
+        ${PROJECT_SOURCES}
+    )
+# Define target properties for Android with Qt 6 as:
+#    set_property(TARGET amcset_gui APPEND PROPERTY QT_ANDROID_PACKAGE_SOURCE_DIR
+#                 ${CMAKE_CURRENT_SOURCE_DIR}/android)
+# For more information, see https://doc.qt.io/qt-6/qt-add-executable.html#target-creation
 
-target_link_libraries(amcset_gui PRIVATE Qt6::Widgets compiler_flags)
+    qt_create_translation(QM_FILES ${CMAKE_SOURCE_DIR} ${TS_FILES})
+else()
+    if(ANDROID)
+        add_library(amcset_gui SHARED
+            ${PROJECT_SOURCES}
+        )
+# Define properties for Android with Qt 5 after find_package() calls as:
+#    set(ANDROID_PACKAGE_SOURCE_DIR "${CMAKE_CURRENT_SOURCE_DIR}/android")
+    else()
+        add_executable(amcset_gui
+            ${PROJECT_SOURCES}
+        )
+    endif()
 
+    qt5_create_translation(QM_FILES ${CMAKE_SOURCE_DIR} ${TS_FILES})
+endif()
+
+target_link_libraries(amcset_gui PRIVATE Qt${QT_VERSION_MAJOR}::Widgets)
+
+# Qt for iOS sets MACOSX_BUNDLE_GUI_IDENTIFIER automatically since Qt 6.1.
+# If you are developing for iOS or macOS you should consider setting an
+# explicit, fixed bundle identifier manually though.
+if(${QT_VERSION} VERSION_LESS 6.1.0)
+  set(BUNDLE_ID_OPTION MACOSX_BUNDLE_GUI_IDENTIFIER com.example.amcset_gui)
+endif()
 set_target_properties(amcset_gui PROPERTIES
+    ${BUNDLE_ID_OPTION}
+    MACOSX_BUNDLE_BUNDLE_VERSION ${PROJECT_VERSION}
+    MACOSX_BUNDLE_SHORT_VERSION_STRING ${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}
     MACOSX_BUNDLE TRUE
     WIN32_EXECUTABLE TRUE
 )
+
+include(GNUInstallDirs)
+install(TARGETS amcset_gui
+    BUNDLE DESTINATION .
+    LIBRARY DESTINATION ${CMAKE_INSTALL_LIBDIR}
+    RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR}
+)
+
+if(QT_VERSION_MAJOR EQUAL 6)
+    qt_finalize_executable(amcset_gui)
+endif()
