+++ amcset_gui/mainwindow.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>amcset_main_window</class>
- <widget class="QMainWindow" name="amcset_main_window">
+ <class>MainWindow</class>
+ <widget class="QMainWindow" name="MainWindow">
   <property name="geometry">
    <rect>
     <x>0</x>
@@ -11,41 +11,31 @@
    </rect>
   </property>
   <property name="windowTitle">
-   <string>amcset_main_window</string>
+   <string>MainWindow</string>
   </property>
   <widget class="QWidget" name="centralwidget">
    <widget class="QSplitter" name="splitter">
     <property name="geometry">
      <rect>
-      <x>250</x>
-      <y>110</y>
+      <x>240</x>
+      <y>140</y>
       <width>256</width>
-      <height>217</height>
+      <height>240</height>
      </rect>
     </property>
     <property name="orientation">
      <enum>Qt::Vertical</enum>
     </property>
-    <widget class="QWidget" name="">
-     <layout class="QVBoxLayout" name="verticalLayout">
-      <item>
-       <widget class="QPushButton" name="pushButton">
-        <property name="text">
-         <string>Prompt Server</string>
-        </property>
-       </widget>
-      </item>
-      <item>
-       <widget class="QTextEdit" name="textEdit"/>
-      </item>
-      <item>
-       <widget class="QPushButton" name="pushButton_2">
-        <property name="text">
-         <string>Exit</string>
-        </property>
-       </widget>
-      </item>
-     </layout>
+    <widget class="QPushButton" name="promptServer">
+     <property name="text">
+      <string>Prompt Server</string>
+     </property>
+    </widget>
+    <widget class="QTextEdit" name="textEdit"/>
+    <widget class="QPushButton" name="exitButton">
+     <property name="text">
+      <string>Exit</string>
+     </property>
     </widget>
    </widget>
   </widget>
